new server_adapters logic
